midterm nav and grid layout
I have formed a nav but I am having trouble with centering an creating a mobile view menu.

The grids I have put into the css code are also giving me trouble; they are not showing up on the page. I hope the layout is able to show as I think they are strong and are design oriented.